Move GeneralAnnotatedTypeFactory to inner class.
This should prevent new uses of this class.
